requirement.txt
python==3.7.10
pytorch==1.7.1

matplotlib==3.4.2
scipy==1.6.3
tqdm==4.59.0